Implement offline backup, and remove the backup facility from the web pages.
It's quite a slow operation, and could cause the server framework to
time out.

We'll do it on a cron job instead, which is also more reliable, and
will put the resulting files on the server, where they can be as
secure as the database itself.  Some other activity can then copy them
off, in case of losing the server.